A function to calculate area of extent change and generate change rasters for reporting.

extent_change calculates change between all consecutive vegetation classification rasters and outputs summary area stats and change shape files.

Usage

extent_change(irast, areaname, iregions, attribname, ext = ".tif")

Arguments

irast

Character file path to input veg class rasters that have been through the veg classification process, i.e. those that are found in `veg_class\`.

areaname

Character vector representing the geographical area that the user is processing, e.g. marine park acronym. It will be used for inclusion to the output name.

iregions

Character file path to a shape file (including extension) that defines reporting regions. The shape file should have an attribute column that defines the overall reporting "region" plus "site", if applicable, such as "NatPark_site1", "NatPark_site2" etc. The underscore delineates region from site.

attribname

Character string of the name of the attribute column that contains the region information.

ext

Character representation of the file extension of the input rasters. Defaults to ".tif" as this is the preferred file type.

Value

The csv and shape files are exported to `extent_change\`.

Details

This function calculates change between all consecutive vegetation classification rasters created from running veg_class. It then outputs extent change area statistics as well as extent change shape files to spatially show where change has occurred.

Author

Bart Huntley, bart.huntley@dbca.wa.gov.au

Examples

if (FALSE) { # \dontrun{
extent_change(irast = "veg_class", areaname = "lgcsmp", iregions = "vectors/regions.shp",
    attribname = "regions")
} # }