A function for calculating trend and trend class images

trend_class creates trend and trend class rasters from user specified time periods.

trend_class(irast, rastkey, end, period, classes)

Arguments

irast
Character file path to input veg density rasters that have been masked and found in either `veg_dens_mskd\` or `veg_dens_mskd_cld\`.

rastkey
Character representation of unique string to match in intended rasters. This aids in not selecting additional files generated by other software (e.g. ArcMap).

end
Numeric denoting final year of analysis.

period
Numeric denoting intended length of trend analysis. Note that periods are matched to class bins so ensure that appropriate period length is entered otherwise results will be meaningless.

classes
Character representation of the name of the trend class csv file including file path.

Value

A raster of slope values and a raster of trend classes are written to `trend_class\` for the intended period of analysis.

Details

Given an end year and a period length this function stacks the appropriate annual rasters and outputs a linear trend raster and a trend class raster based on a user supplied classification in csv format. The csv must have the following three column format so that each row forms a bin:

lower - numerical low value for bin

upper - numerical high value (<=) for bin

reclass - numerical classification value

This classification is based on binning the slope values from the regression and is therefore sensitive to the period length. Ensure that the classification matches the intended period of analysis.

Note to accommodate any missing annual rasters, due to for example availability of suitable imagery, the function will create "dummy" rasters to pad the stack accordingly. These "dummy" rasters ensure the time variable is handled correctly in the regression. These "dummies" are deleted as the function cleans up after itself.

Author

Bart Huntley, bart.huntley@dbca.wa.gov.au

Examples

if (FALSE) {
trend_class(irast = "./veg_dens_mskd_for_trends", rastkey = ".tif",
    end = 2009, period = 10, classes = "./supplementary/trend_classes.csv")
}