A function to perform vegetation classifications

veg_class performs a raster reclassification based on a user supplied set of meaningful bins.

veg_class(irast, rastkey, imask, maskkey, classes)

Arguments

irast

Character file path to input veg density rasters that have been through the masking process, i.e. those that are found in `veg_dens_mskd\` or `veg_dens_mskd_cld\`.

rastkey

Character representation of unique string to match in intended rasters. This aids in not selecting additional files generated by other software (e.g. ArcMap).

imask

A character file path to directory of raster masks that were created using a "cloud_mask" function if a "probable" classification is desired. If there were none or you would prefer that no "probable" classifications be carried out use FALSE.

maskkey

Character representation of the file extension of the raster masks if a "probable" classification is desired. If not use NA

classes

Character representation of the name of the density classes csv file including file path.

Value

All input rasters will be reclassified and be written to either `veg_class_cloud_prob/` for "probable" classifications or `veg_class/` if standard classifications are carried out.

Details

This function is designed to take the continuous values, in say a raster of vegetation densities, and convert them to numeric classes based on some prior determination. The user must supply a csv of density classes with the following three column format so that each row forms a bin:

lower - numerical low value for bin

upper - numerical high value (<=) for bin

reclass - numerical classification value

The numerical classification can take any single integer value except 6. The integer 6 is reserved for any values that had a masked value of -99 created by using the "cloud_mask" functions.

The function can also perform some "probable" value classification in areas that are masked with a -99 value. It will identify the same pixels in the previous year's raster and return those classified values plus 10. The reasoning here is that those pixels possibly would not have changed density but they can still be identified and removed (e.g. any value > 10) in downstream analysis.

Author

Bart Huntley, bart.huntley@dbca.wa.gov.au

Examples

if (FALSE) {
veg_class(irast = "./veg_dens_mskd", rastkey = ".tif",
    imask = FALSE, maskkey = NA,
    classes = "./supplementary/density_classes.csv")
}