A function to turn vegetation index raster data into a vegetation density raster for further analysis.

veg_dens takes an annual series of vegetation index raster mosaics and converts them to a vegetation density (cover) product.

Usage

veg_dens(
  irast,
  areaname,
  ext = ".tif",
  calibration = "supplementary/calibration.csv"
)

Arguments

irast

Character file path to input vegetation index rasters.

areaname

Character vector representing the geographical area that the user is processing. It is good practice to add in the satellite sensor used to create the index images to help keep track of the source. Just do not add any numbers to the names. An example might be "lgcsmp_lsat" or "lgcsmp_sent" for Lalang-garram landsat and Lalang-garram sentinel respectively. It will be used for inclusion to the output csv name.

ext

Character representation of the input file type. Defaults to ".tif" as this is the preferred file type.

calibration

Character representation of the name of the calibration csv file including file path. Defaults to "./supplementary/calibration.csv" which works with the suggested project folder structure and workflow.

Value

For each input raster a vegetation density raster of the same will be written to file in a folder named `veg_dens/`.

Details

Density to vegetation index relationship must be established through prior analysis and the calibration file will provide the coefficients from this. The calibration file is a simple csv with a value in 5 columns named as follows:

coef - the coefficient of the regression

intercept - the intercept of the regression

multiple - a value to multiply by to bring output to a percentage

lower - lower limit of acceptable vegetation density

upper - upper limit of acceptable vegetation density

All input raster mosaics should have the same extents and cell size. Whilst it won't affect the performance of this function, downstream processing will fail. See package vignette for full explanation of processing work flow.

Input raster mosaics can follow any naming convention that makes sense to the project but must contain a 4 digit representation of the year and contain NO other numerals. An example might be "lgcsmp_ndvi_2023.tif".

Author

Bart Huntley, bart.huntley@dbca.wa.gov.au

Examples

if (FALSE) { # \dontrun{
veg_dens(irast = "Z:/DEC/projectXX/mosaics", areaname = "lgcsmp_lsat")
    } # }